Split Editor and Pedal component in separate source files
